LayoutTests/storage/indexeddb/intversion-abort-in-initial-upgradeneeded-expected.txt

Test that an abort in the initial upgradeneeded sets version back to 0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-abort-in-initial-upgradeneeded.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname, 2)
request.onupgradeneeded = upgradeNeeded
FIXME: This should get an error event of type AbortError
request.onerror = unexpectedErrorCallback

upgradeNeeded():
PASS db.version is 2
transaction.abort()

onAbort():
FAIL event.target.db.version should be 0. Was 2.
PASS successfullyParsed is true

TEST COMPLETE
